LPS-168155 Move MediaWikiImporter and its related usages to wiki-service internal package.
