+++ modules/apps/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/importer/MediaWikiImporter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.web.internal.importer;
+package com.liferay.wiki.internal.importer;
 
 import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.service.AssetTagLocalService;
@@ -55,11 +55,11 @@ import com.liferay.wiki.constants.WikiWebKeys;
 import com.liferay.wiki.exception.ImportFilesException;
 import com.liferay.wiki.exception.NoSuchPageException;
 import com.liferay.wiki.importer.WikiImporter;
+import com.liferay.wiki.internal.translator.MediaWikiToCreoleTranslator;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
 import com.liferay.wiki.service.WikiPageLocalService;
 import com.liferay.wiki.validator.WikiPageTitleValidator;
-import com.liferay.wiki.web.internal.translator.MediaWikiToCreoleTranslator;
 
 import java.io.IOException;
 import java.io.InputStream;
