+++ modules/apps/wiki/wiki-service/src/test/java/com/liferay/wiki/internal/translator/ClassicToCreoleTranslatorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.web.internal.translator;
+package com.liferay.wiki.internal.translator;
 
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
