+++ modules/apps/wiki/wiki-service/src/test/java/com/liferay/wiki/internal/translator/MediaWikiToCreoleTranslatorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.web.internal.translator;
+package com.liferay.wiki.internal.translator;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
